+++ module/data/_module.mjs
@@ -17,3 +17,4 @@ export {default as ArkhamHorrorTome} from "./item-tome.mjs";
 export {default as ArkhamHorrorRelic} from "./item-relic.mjs";
 export {default as ArkhamHorrorInjury} from "./item-injury.mjs";
 export {default as ArkhamHorrorTrauma} from "./item-trauma.mjs";
+export {default as ArkhamHorrorFavor} from "./item-favor.mjs";